Scrap total supple and reserve for atokens

+++ pkg/dia/Messages.go
@@ -428,7 +428,7 @@ type SynthAssetSupply struct {
 	LockedUnderlying float64 // Amount of underlying asset locked in the contract.
 	NumMint          int64   // Total number of synth asset mint events (optional).
 	NumRedeem        int64   // Total number of underlying asset redeem events (optional).
-	BlockNumber      uint8
+	BlockNumber      uint64
 }
 
 type TradesBlockData struct {
